refactor: changed frame advantages from i8 to i32

+++ ggrs/src/network/network_stats.rs
@@ -13,9 +13,9 @@ pub struct NetworkStats {
     /// The number of frames GGRS calculates that the local client is behind the remote client at this instant in time.  
     /// For example, if at this instant the current game client is running frame 1002 and the remote game client is running frame 1009,
     /// this value will mostly likely roughly equal 7.
-    pub local_frames_behind: i8,
+    pub local_frames_behind: i32,
     /// The same as `local_frames_behind`, but calculated from the perspective of the remote player.
-    pub remote_frames_behind: i8,
+    pub remote_frames_behind: i32,
 }
 
 impl NetworkStats {
